+++ Sources/TuistKit/Generator/ProjectGenerator.swift
@@ -104,6 +104,10 @@ final class ProjectGenerator: ProjectGenerating {
                                                 options: options,
                                                 graph: graph)
 
+        generateTestTargetIdentity(project: project,
+                                   pbxproj: pbxproj,
+                                   pbxProject: pbxProject)
+
         return try write(xcodeprojPath: xcodeprojPath,
                          nativeTargets: nativeTargets,
                          workspace: workspace,
@@ -130,6 +134,7 @@ final class ProjectGenerator: ProjectGenerating {
                                     projects: [],
                                     projectRoots: [],
                                     targets: [])
+
         pbxproj.add(object: pbxProject)
         pbxproj.rootObject = pbxProject
         return pbxProject
@@ -166,6 +171,38 @@ final class ProjectGenerator: ProjectGenerating {
         return nativeTargets
     }
 
+    fileprivate func generateTestTargetIdentity(project _: Project,
+                                                pbxproj: PBXProj,
+                                                pbxProject: PBXProject) {
+        func testTargetName(_ target: PBXTarget) -> String? {
+            guard let buildConfigurations = target.buildConfigurationList?.buildConfigurations else {
+                return nil
+            }
+
+            return buildConfigurations
+                .compactMap { $0.buildSettings["TEST_TARGET_NAME"] as? String }
+                .first
+        }
+
+        let testTargets = pbxproj.nativeTargets.filter { $0.productType == .uiTestBundle || $0.productType == .unitTestBundle }
+
+        for testTarget in testTargets {
+            guard let name = testTargetName(testTarget) else {
+                continue
+            }
+
+            guard let target = pbxproj.targets(named: name).first else {
+                continue
+            }
+
+            var attributes = pbxProject.targetAttributes[testTarget] ?? [:]
+
+            attributes["TestTargetID"] = target
+
+            pbxProject.setTargetAttributes(attributes, target: testTarget)
+        }
+    }
+
     fileprivate func write(xcodeprojPath: AbsolutePath,
                            nativeTargets: [String: PBXNativeTarget],
                            workspace: XCWorkspace,
@@ -179,7 +216,8 @@ final class ProjectGenerator: ProjectGenerating {
             try writeXcodeproj(workspace: workspace,
                                pbxproj: pbxproj,
                                xcodeprojPath: temporaryPath)
-            generatedProject = GeneratedProject(path: temporaryPath,
+            generatedProject = GeneratedProject(pbxproj: pbxproj,
+                                                path: temporaryPath,
                                                 targets: nativeTargets,
                                                 name: xcodeprojPath.components.last!)
             try writeSchemes(project: project,
@@ -188,7 +226,7 @@ final class ProjectGenerator: ProjectGenerating {
             try fileHandler.replace(xcodeprojPath, with: temporaryPath)
         }
 
-        return generatedProject.at(path: xcodeprojPath)
+        return try generatedProject.at(path: xcodeprojPath)
     }
 
     fileprivate func writeXcodeproj(workspace: XCWorkspace,