+++ fixtures/app_with_framework_and_tests/App/AppDelegate.swift
@@ -14,4 +14,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         window?.makeKeyAndVisible()
         return true
     }
+    
+    func hello() -> String {
+        return "AppDelegate.hello()"
+    }
+    
 }