+++ fixtures/ios_app_with_setup/Tests/AppTests.swift
@@ -3,4 +3,12 @@ import XCTest
 
 @testable import App
 
-final class AppTests: XCTestCase {}
+final class AppTests: XCTestCase {
+    
+    func testHello() {
+        let sut = AppDelegate()
+        
+        XCTAssertEqual("AppDelegate.hello()", sut.hello())
+    }
+    
+}